**`[Breaking Change]`** Moved upload image processing methods like ScaleImage etc. in UploadStorageExtensions file to DefaultUploadProcessor as virtual methods. Removed FileUploadBehavior.CheckUploadedImageAndCreateThumbs and used existing IUploadProcessor interface instead. Set additional metadata like EntityType (row full name), EntityId (value of id property), EntityProperty (the property that the upload editor is placed on) so that files can be traced back to their original entities. Also set generated ImageSize and ThumbSize as metadata to avoid loading image files to get actual sizes when required. This is also a preparation for adding custom metadata like tags, description etc. to some files if desired.

+++ src/Serenity.Net.Services/Upload/DefaultUploadProcessor.cs
@@ -1,4 +1,3 @@
-
 using Microsoft.Extensions.Logging;
 using System.IO;
 
@@ -20,14 +19,33 @@ public class DefaultUploadProcessor(IImageProcessor imageProcessor, IUploadStora
     ILogger<DefaultUploadProcessor> logger = null,
     IUploadAVScanner avScanner = null) : IUploadProcessor
 {
-    private readonly IImageProcessor imageProcessor = imageProcessor ?? throw new ArgumentNullException(nameof(imageProcessor));
-    private readonly IUploadStorage uploadStorage = uploadStorage ?? throw new ArgumentNullException(nameof(uploadStorage));
-    private readonly IUploadValidator uploadValidator = uploadValidator ?? throw new ArgumentNullException(nameof(uploadValidator));
-    private readonly ILogger<DefaultUploadProcessor> logger = logger;
-    private readonly IUploadAVScanner avScanner = avScanner;
+    /// <summary>
+    /// Image processor
+    /// </summary>
+    protected readonly IImageProcessor imageProcessor = imageProcessor ?? throw new ArgumentNullException(nameof(imageProcessor));
+
+    /// <summary>
+    /// Upload storage
+    /// </summary>
+    protected readonly IUploadStorage uploadStorage = uploadStorage ?? throw new ArgumentNullException(nameof(uploadStorage));
+
+    /// <summary>
+    /// Upload validator
+    /// </summary>
+    protected readonly IUploadValidator uploadValidator = uploadValidator ?? throw new ArgumentNullException(nameof(uploadValidator));
+
+    /// <summary>
+    /// Logger
+    /// </summary>
+    protected readonly ILogger<DefaultUploadProcessor> logger = logger;
+
+    /// <summary>
+    /// AV Scanner
+    /// </summary>
+    protected readonly IUploadAVScanner avScanner = avScanner;
 
     /// <inheritdoc/>
-    public ProcessedUploadInfo Process(Stream stream, string filename, IUploadOptions options)
+    public virtual ProcessedUploadInfo Process(Stream stream, string filename, IUploadOptions options)
     {
         if (stream is null)
             throw new ArgumentNullException(nameof(stream));
@@ -38,13 +56,20 @@ public class DefaultUploadProcessor(IImageProcessor imageProcessor, IUploadStora
         if (options is null)
             throw new ArgumentNullException(nameof(options));
 
-        avScanner?.Scan(stream, filename);
+        UploadPathHelper.CheckFileNameSecurity(filename);
+
+        bool isExistingTemporary = filename.StartsWith("temporary/", StringComparison.Ordinal) &&
+            uploadStorage.FileExists(filename);
+
+        if (!isExistingTemporary)
+            avScanner?.Scan(stream, filename);
 
         var result = new ProcessedUploadInfo
         {
             FileSize = stream.Length
         };
 
+        var success = false;
         try
         {
             try
@@ -60,22 +85,25 @@ public class DefaultUploadProcessor(IImageProcessor imageProcessor, IUploadStora
                 {
                     uploadStorage.PurgeTemporaryFiles();
 
-                    var basePath = "temporary/" + Guid.NewGuid().ToString("N");
-                    stream.Seek(0, SeekOrigin.Begin);
-                    result.TemporaryFile = uploadStorage.WriteFile(basePath + Path.GetExtension(filename), 
-                        stream, OverwriteOption.Disallowed);
+                    if (!isExistingTemporary)
+                    {
+                        var basePath = "temporary/" + Guid.NewGuid().ToString("N");
+                        stream.Seek(0, SeekOrigin.Begin);
+                        result.TemporaryFile = uploadStorage.WriteFile(basePath + Path.GetExtension(filename),
+                            stream, OverwriteOption.Disallowed);
+                    }
+                    else
+                    {
+                        result.TemporaryFile = filename;
+                    }
+
                     result.IsImage = isImageExtension && image != null;
                     if (result.IsImage)
                     {
-                        var (width, height) = imageProcessor.GetImageSize(image);
-                        result.ImageWidth = width;
-                        result.ImageHeight = height;
                         stream.Close();
-                        result.TemporaryFile = UploadStorageExtensions.ScaleImageAndCreateAllThumbs(image, imageProcessor,
-                            options as IUploadImageOptions ?? new UploadOptions(),
-                            uploadStorage, result.TemporaryFile, OverwriteOption.Overwrite);
+                        result.TemporaryFile = ProcessImage(image, options as IUploadImageOptions ?? new UploadOptions(), result.TemporaryFile);
                     }
-                    result.Success = true;
+                    success = true;
                 }
                 finally
                 {
@@ -84,19 +112,225 @@ public class DefaultUploadProcessor(IImageProcessor imageProcessor, IUploadStora
             }
             catch (Exception ex)
             {
-                result.ErrorMessage = ex.Message;
-                result.Success = false;
                 logger?.LogError(ex, "Error occurred while processing upload!");
+                throw;
             }
         }
         finally
         {
-            if (!result.Success && !string.IsNullOrEmpty(result.TemporaryFile))
+            if (!success && !string.IsNullOrEmpty(result.TemporaryFile) && !isExistingTemporary)
                 uploadStorage.DeleteFile(result.TemporaryFile);
 
             stream?.Dispose();
         }
 
         return result;
+    }
+
+    /// <summary>
+    /// Depending on the image upload options, scales image, creates default and
+    /// additional thumbs and saves them to the upload storage files.
+    /// </summary>
+    /// <param name="image">Input image</param>
+    /// <param name="options">Upload image options</param>
+    /// <param name="imageFile">Image file path</param>
+    /// <returns>Scaled image file path. It might be different then the passed file if scaling is performed.
+    /// E.g. the uploaded file might be .png while the scaled image might be .jpg</returns>
+    protected virtual string ProcessImage(object image, IUploadImageOptions options, string imageFile)
+    {
+        imageFile = ScaleMainImage(image, options, imageFile)?.Filename ?? imageFile;
+        CreateDefaultThumb(image, options, imageFile);
+        CreateAdditionalThumbs(image, options, imageFile);
+        return imageFile;
+    }
+
+    /// <summary>
+    /// Scales the temporary image with provided upload image options if required 
+    /// based on the options and saves the result to the target upload storage file
+    /// </summary>
+    /// <param name="image">Image object</param>
+    /// <param name="options">Image upload options</param>
+    /// <param name="imageFile">Main image file</param>
+    /// <returns>The resulting image file path</returns>
+    /// <exception cref="ArgumentNullException">image or options is null</exception>
+    protected virtual ScaleImageAsResult ScaleMainImage(object image, IUploadImageOptions options, string imageFile)
+    {
+        if (image is null)
+            throw new ArgumentNullException(nameof(image));
+
+        if (options is null)
+            throw new ArgumentNullException(nameof(options));
+
+        if (string.IsNullOrEmpty(imageFile))
+            throw new ArgumentNullException(nameof(imageFile));
+
+        var (imageWidth, imageHeight) = imageProcessor.GetImageSize(image);
+        var scaleSmaller = options.ScaleSmaller == true;
+
+        ScaleImageAsResult result = null;
+        if ((options.ScaleWidth > 0 || options.ScaleHeight > 0) &&
+            (options.ScaleWidth != imageWidth || options.ScaleHeight != imageHeight) &&
+            ((options.ScaleWidth > 0 && (scaleSmaller || options.ScaleWidth < imageWidth)) ||
+             (options.ScaleHeight > 0 && (scaleSmaller || options.ScaleHeight < imageHeight))))
+        {
+            var baseFile = Path.ChangeExtension(imageFile, null);
+            var originalName = uploadStorage.GetOriginalName(imageFile);
+            result = ScaleImageAs(image, options.ScaleWidth, options.ScaleHeight, options.ScaleMode, 
+                options.ScaleBackColor, new ImageEncoderParams { Quality = options.ScaleQuality },
+                baseFile + ".jpg");
+
+            imageWidth = result.Width;
+            imageHeight = result.Height;
+
+            if (!string.IsNullOrEmpty(originalName))
+                uploadStorage.SetOriginalName(result.Filename, Path.ChangeExtension(originalName, ".jpg"));
+        }
+
+        uploadStorage.SetFileMetadata(result?.Filename ?? imageFile, new Dictionary<string, string>()
+        {
+            [FileMetadataKeys.ImageSize] = imageWidth.ToInvariant() + "x" + imageHeight.ToInvariant(),
+        }, overwriteAll: false);
+
+        return result;
+    }
+
+    /// <summary>
+    /// Creates the default thumbnail for image if the size is provided
+    /// in the upload image options (ThumbWidth and ThumbHeight >= 0) and saves it to the target upload storage file
+    /// </summary>
+    /// <param name="image">Image</param>
+    /// <param name="options">Upload image options</param>
+    /// <param name="imageFile">Main image file</param>
+    /// <exception cref="ArgumentNullException">image, options or temporaryFile is null</exception>
+    public virtual ScaleImageAsResult CreateDefaultThumb(object image, IUploadImageOptions options, string imageFile)
+    {
+        if (image is null)
+            throw new ArgumentNullException(nameof(image));
+
+        if (options is null)
+            throw new ArgumentNullException(nameof(options));
+
+        if ((options.ThumbWidth > 0 || options.ThumbHeight > 0) &&
+            (options.ThumbWidth >= 0 && options.ThumbHeight >= 0))
+        {
+            var result = ScaleImageAs(image, options.ThumbWidth, options.ThumbHeight, 
+                options.ThumbMode, options.ThumbBackColor, 
+                new ImageEncoderParams { Quality = options.ThumbQuality },
+                Path.ChangeExtension(imageFile, null) + "_t.jpg");
+
+            uploadStorage.SetFileMetadata(imageFile, new Dictionary<string, string>()
+            {
+                [FileMetadataKeys.ThumbSize] = result.Width.ToInvariant() + "x" + result.Height.ToInvariant()
+            }, overwriteAll: false);
+
+            return result;
+        }
+
+        return null;
+    }
+
+    /// <summary>
+    /// Creates additional thumbs if specified in the upload image options,
+    /// and saves them to the target upload storage
+    /// </summary>
+    /// <param name="image">Image</param>
+    /// <param name="options">Upload image options</param>
+    /// <param name="imageFile">Main image file</param>
+    /// <exception cref="ArgumentNullException">image, options or temporaryFile is null</exception>
+    /// <exception cref="ArgumentOutOfRangeException">options.ThumbSizes contains invalid values</exception>
+    protected virtual IEnumerable<ScaleImageAsResult> CreateAdditionalThumbs(object image, 
+        IUploadImageOptions options, string imageFile)
+    {
+        if (image is null)
+            throw new ArgumentNullException(nameof(image));
+
+        if (options is null)
+            throw new ArgumentNullException(nameof(options));
+
+        if (string.IsNullOrEmpty(imageFile))
+            throw new ArgumentNullException(nameof(imageFile));
+
+        var result = new List<ScaleImageAsResult>();
+        var thumbSizes = options.ThumbSizes.TrimToNull();
+        if (thumbSizes is null)
+            return result;
+
+        var baseFile = Path.ChangeExtension(imageFile, null);
+
+        foreach (var sizeStr in thumbSizes.Replace(";", ",", StringComparison.Ordinal).Split([',']))
+        {
+            var dims = sizeStr.ToUpperInvariant().Split(['X']);
+            if (dims.Length != 2 ||
+                !int.TryParse(dims[0], out int w) ||
+                !int.TryParse(dims[1], out int h) ||
+                w < 0 ||
+                h < 0 ||
+                (w == 0 && h == 0))
+                throw new ArgumentOutOfRangeException(nameof(thumbSizes));
+
+            var thumbFile = baseFile + "_t" + w.ToInvariant() + "x" + h.ToInvariant() + ".jpg";
+
+            result.Add(ScaleImageAs(image, w, h, options.ThumbMode, options.ThumbBackColor,
+                new ImageEncoderParams { Quality = options.ThumbQuality },
+                thumbFile));
+        }
+
+        return result;
+    }
+
+    /// <summary>
+    /// Scales an image and saves it to an upload storage file
+    /// </summary>
+    /// <param name="image">Source image</param>
+    /// <param name="width">Target width</param>
+    /// <param name="height">Target height</param>
+    /// <param name="mode">Scale mode</param>
+    /// <param name="backgroundColor">Pad color</param>
+    /// <param name="encoderParams">Encoder parameters for target image</param>
+    /// <param name="targetFile">Path</param>
+    /// <exception cref="ArgumentNullException">One of inputs is null</exception>
+    protected virtual ScaleImageAsResult ScaleImageAs(object image, int width, int height, 
+        ImageScaleMode mode, string backgroundColor, ImageEncoderParams encoderParams, string targetFile)
+    {
+        var scaledImage = imageProcessor.Scale(image, width, height, mode, backgroundColor, inplace: false);
+        try
+        {
+            var size = imageProcessor.GetImageSize(scaledImage);
+
+            using var ms = new MemoryStream();
+            imageProcessor.Save(scaledImage, ms, "image/jpeg", encoderParams);
+            ms.Seek(0, SeekOrigin.Begin);
+            return new ScaleImageAsResult
+            {
+                Filename = uploadStorage.WriteFile(targetFile, ms, OverwriteOption.Overwrite),
+                Width = size.width,
+                Height = size.height
+            };
+        }
+        finally
+        {
+            (scaledImage as IDisposable)?.Dispose();
+        }
+    }
+
+    /// <summary>
+    /// Result for the scale image as operation
+    /// </summary>
+    public class ScaleImageAsResult
+    {
+        /// <summary>
+        /// Resulting filename
+        /// </summary>
+        public string Filename { get; set; }
+
+        /// <summary>
+        /// Resulting image height
+        /// </summary>
+        public int Height { get; set; }
+
+        /// <summary>
+        /// Resulting image width
+        /// </summary>
+        public int Width { get; set; }
     }
 }
